+++ src/index.ts
@@ -18,5 +18,6 @@ export * from './utilities';
 export * from './errors/InvalidResourceError';
 export * from './errors/ResourceNotFoundError';
 export * from './errors/ResourceVersionNotFoundError';
+export * from './errors/UnauthorizedError';
 export * from './errors/TooManyConcurrentExportRequestsError';
 export { stubs } from './stubs';