+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.0.0-beta.1] - 2020-11-04
+
+- feat: export UnauthorizedError to be usable
+
+## [3.0.0-beta.0] - 2020-11-04
+
+- feat: Add required methods for SMART on FHIR interactions
+
 ## [2.0.0] - 2020-09-25
 
 - feat: Added `getAllowedResourceTypesForOperation` method to `Authorization` interface.
 - feat: Updated `Authorization.isAuthorized` to return a promise.
-
 
 ## [1.0.0] - 2020-08-31